Fix potential infinite generator in `Market.trades`.
Since 5924ac129721f432f3e44d8390dba0c895014ab0 was added,
`Market.trades` is a generator and not a function. In it,
there is a single exit condition -- `cnt >= limit`. However,
this condition might never be reached for situations, where
the number of returned results is less than required limit.

The simplest test to verify this is to open an empty market
 -- `trades` will never return. Similar thing could happen
if you ask for, eg, 50 trades, but the start - stop interval
only has 30.