Always add final padding during memo encryption.
This avoids weird edge cases, when the message is exactly the length
of 16 byte span, and the last byte then gets erroneously treated
as padding marker.

Note: messages with length of 15 (or 16+15, etc) bytes still
get encoded as before, it's only when the edge is pushed
and there's no place for a "padding marker", we add 16 more
bytes (yes, they all consist of the same 0x0F byte).

Example:

"1234567890A"  = "1234567890A" +  ("\x0F" * 1)
"1234567890AA" = "1234567890AA" + ("\x0F" * 16)

Note: bitcoin-core does it the same way, which makes sense.
(As it's the only sane way to distinguish between actual last byte
of the message and padding markers.)